refactor(app): isolate DB wedge recovery
