+++ apps/screenpipe-app-tauri/src-tauri/src/recording.rs
@@ -122,244 +122,10 @@ pub fn local_api_context_from_app(app: &tauri::AppHandle) -> LocalApiContext {
 const RESTART_COOLDOWN_SECS: u64 = 30;
 const CAPTURE_RESTART_MEETING_REATTACH_WINDOW: Duration = Duration::from_secs(120);
 
-/// Shared state for the DB-wedge auto-recovery circuit breaker.
-#[derive(Default)]
-pub struct DbWedgeState {
-    /// Timestamps of recent auto-restarts, so a DB that stays broken after a
-    /// restart (genuine on-disk corruption, which a restart can't repair)
-    /// cannot restart-storm.
-    restarts: std::collections::VecDeque<std::time::Instant>,
-    /// Whether the user has already been told auto-recovery gave up this
-    /// episode. The persistent-failure hook can keep firing while the breaker
-    /// is tripped, so this dedupes the "needs recovery" notification.
-    gave_up_notified: bool,
-}
-
-pub type DbWedgeBreaker = Arc<std::sync::Mutex<DbWedgeState>>;
-
-pub fn new_db_wedge_breaker() -> DbWedgeBreaker {
-    Arc::new(std::sync::Mutex::new(DbWedgeState::default()))
-}
-
-/// What the circuit breaker decided to do about one persistent-failure signal.
-#[derive(Debug, PartialEq, Eq)]
-enum WedgeAction {
-    /// Attempt a stop→spawn restart (the timestamp was recorded).
-    Restart,
-    /// Too many restarts in the window — don't restart. `notify` is true only
-    /// the first time we give up this episode, so a hook that keeps firing
-    /// while the breaker is tripped doesn't spam the notification panel.
-    GiveUp { notify: bool },
-}
-
-impl DbWedgeState {
-    /// Age out restart timestamps older than `window`, then decide whether to
-    /// restart again. On `Restart` the new attempt is recorded and the give-up
-    /// notice is re-armed for this episode.
-    fn decide(
-        &mut self,
-        now: std::time::Instant,
-        window: Duration,
-        max_restarts: usize,
-    ) -> WedgeAction {
-        while self
-            .restarts
-            .front()
-            .is_some_and(|t| now.duration_since(*t) > window)
-        {
-            self.restarts.pop_front();
-        }
-        if self.restarts.len() >= max_restarts {
-            let notify = !self.gave_up_notified;
-            self.gave_up_notified = true;
-            return WedgeAction::GiveUp { notify };
-        }
-        self.restarts.push_back(now);
-        self.gave_up_notified = false;
-        WedgeAction::Restart
-    }
-}
-
-/// Max auto-restarts allowed inside `DB_WEDGE_BREAKER_WINDOW` before giving up.
-const DB_WEDGE_MAX_RESTARTS: usize = 3;
-const DB_WEDGE_BREAKER_WINDOW: Duration = Duration::from_secs(600);
-/// Coalesce a burst of persistent-failure signals before acting.
-const DB_WEDGE_DEBOUNCE: Duration = Duration::from_secs(15);
-
-/// Result of revalidating a persistent-failure signal after the debounce.
-#[derive(Debug, PartialEq, Eq)]
-enum DbWedgeRecoveryDecision {
-    Restart,
-    SkipNoServer,
-    SkipSupersededGeneration,
-    SkipRecovered,
-}
-
-fn db_wedge_recovery_decision(
-    signaled_health: &screenpipe_db::WriteQueueHealth,
-    signaled_recovery_epoch: u64,
-    current_health: Option<&screenpipe_db::WriteQueueHealth>,
-) -> DbWedgeRecoveryDecision {
-    let Some(current_health) = current_health else {
-        return DbWedgeRecoveryDecision::SkipNoServer;
-    };
-    if !signaled_health.is_same_instance(current_health) {
-        return DbWedgeRecoveryDecision::SkipSupersededGeneration;
-    }
-    if current_health.fatal_run_recovery_epoch() != signaled_recovery_epoch {
-        return DbWedgeRecoveryDecision::SkipRecovered;
-    }
-    DbWedgeRecoveryDecision::Restart
-}
-
-/// Build the `PersistentFailureHook` the DB layer fires when writes wedge
-/// persistently. The hook itself is sync (`Fn()`), so it spawns the async
-/// restart. Captures an `AppHandle` (cheap clone, Send+Sync) and the shared
-/// breaker so restart-storm protection persists across restarts.
-pub fn make_db_wedge_recovery_hook(
-    app: tauri::AppHandle,
-    breaker: DbWedgeBreaker,
-    health: screenpipe_db::WriteQueueHealth,
-) -> screenpipe_db::PersistentFailureHook {
-    std::sync::Arc::new(move || {
-        let app = app.clone();
-        let breaker = breaker.clone();
-        let health = health.clone();
-        let recovery_epoch = health.fatal_run_recovery_epoch();
-        // The hook fires on the dedicated *server* runtime. Recovery removes
-        // that server from state, which intentionally lets its runtime exit;
-        // running this task there would cancel it halfway through respawn.
-        // Dispatch onto Tauri's process-lifetime runtime instead.
-        tauri::async_runtime::spawn(async move {
-            recover_from_db_wedge(app, breaker, health, recovery_epoch).await;
-        });
-    })
-}
-
-async fn recover_from_db_wedge(
-    app: tauri::AppHandle,
-    breaker: DbWedgeBreaker,
-    signaled_health: screenpipe_db::WriteQueueHealth,
-    signaled_recovery_epoch: u64,
-) {
-    // Debounce: let a burst of signals coalesce and any in-flight work settle.
-    tokio::time::sleep(DB_WEDGE_DEBOUNCE).await;
-
-    let recording_state = app.state::<RecordingState>();
-
-    // Serialize the exact-generation claim, teardown, and respawn with every
-    // explicit full stop/start. This closes the debounce TOCTOU where a stale
-    // task could validate server A, a manual restart could install server B,
-    // and the stale task would then tear B down.
-    let _lifecycle_guard = recording_state.server_lifecycle.lock().await;
-
-    // Lock in the documented order and atomically remove only the server
-    // generation that raised this signal. Once it is removed, no other full
-    // lifecycle path can replace it until this recovery releases the outer
-    // lifecycle guard.
-    let mut capture_guard = recording_state.capture.lock().await;
-    let mut server_guard = recording_state.server.lock().await;
-    let current_health = server_guard
-        .as_ref()
-        .map(|core| core.db.write_queue_health());
-    let decision = db_wedge_recovery_decision(
-        &signaled_health,
-        signaled_recovery_epoch,
-        current_health.as_ref(),
-    );
-    if decision != DbWedgeRecoveryDecision::Restart {
-        info!(
-            "db wedge auto-recovery: restart cancelled after debounce ({:?})",
-            decision
-        );
-        return;
-    }
-
-    // Circuit breaker: a DB that stays broken after a restart is on-disk
-    // corruption a restart can't repair, so cap auto-restarts per window.
-    // Decide while the exact generation is still claimed; a skipped stale
-    // signal must not consume restart budget.
-    let action = {
-        let mut state = breaker.lock().unwrap();
-        state.decide(
-            std::time::Instant::now(),
-            DB_WEDGE_BREAKER_WINDOW,
-            DB_WEDGE_MAX_RESTARTS,
-        )
-    };
-    if let WedgeAction::GiveUp { notify } = action {
-        drop(server_guard);
-        drop(capture_guard);
-        error!(
-            "db wedge auto-recovery: {} restarts within {:?} did not clear the write wedge — \
-             in-process restarts can't fix this (poisoned WAL-index pinned by a leaked \
-             connection, or on-disk damage). Surfacing manual recovery.",
-            DB_WEDGE_MAX_RESTARTS, DB_WEDGE_BREAKER_WINDOW
-        );
-        if notify {
-            // In-process restarts are proven futile for this episode. Do not
-            // surprise-relaunch the app on broad DB-shaped errors; surface a
-            // user-visible recovery state instead.
-            crate::db_relaunch::surface_manual_recovery(
-                "db wedge persisted across in-process engine restarts",
-            )
-            .await;
-        }
-        return;
-    }
-
-    let capture = capture_guard.take();
-    let server = server_guard
-        .take()
-        .expect("restart decision requires a current server generation");
-
-    warn!(
-        "db wedge auto-recovery: persistent write failure detected — restarting recording to \
-         rebuild all DB pools + the shared WAL-index"
-    );
-
-    *recording_state.interrupted_meeting.lock().await = None;
-    if let Some(session) = capture {
-        session.stop().await;
-    }
-    server.shutdown().await;
-    // Keep the state guards until shutdown completes. The dedicated server
-    // runtime exits when it can lock `server` and observe None; releasing the
-    // guard earlier can drop that runtime mid-shutdown and cancel the pool/task
-    // cleanup this recovery depends on.
-    drop(server_guard);
-    drop(capture_guard);
-    recording_state.is_starting.store(false, Ordering::SeqCst);
-    recording_state.last_spawn_epoch.store(0, Ordering::SeqCst);
-
-    // The teardown above rebuilds the engine's read/write pools on respawn, but the
-    // secret-store pool is a process-lifetime cache (min_connections=1, no idle
-    // reaping) that would otherwise keep a connection — and the poisoned `-shm`
-    // WAL-index — open across the restart. SQLite only rebuilds `-shm` once the
-    // LAST connection to the db closes, so without this the wedge survives an
-    // in-process restart and recording stays down until a full process exit.
-    // Pools recreate lazily on the next secret access after spawn reopens.
-    screenpipe_secrets::close_all_secret_pools().await;
-
-    // Preserve the latest user capture intent. In particular, stop_capture can
-    // run during the debounce/teardown: the server still needs rebuilding, but
-    // the new server must come back without resurrecting recording.
-    if let Err(e) = spawn_screenpipe_inner(&recording_state, app.clone()).await {
-        // The restart failed to bring the engine back up (e.g. the port never
-        // rebound). Nothing else will retry until the DB layer fires the hook
-        // again — and if the server is fully down it never will — so recording
-        // would otherwise sit silently stopped. Publish on the event bus so the
-        // in-process `db_recovery_notifications` subscriber surfaces it.
-        error!("db wedge auto-recovery: spawn_screenpipe failed: {}", e);
-        let evt = screenpipe_events::DbRecoveryEvent::restart_failed();
-        let _ = screenpipe_events::send_event(evt.event_name(), evt);
-        // A DB-init failure right after a full pool close means the WAL-index
-        // is pinned by something outside our teardown — count it toward the
-        // manual recovery threshold rather than waiting for the health
-        // watchdog to grind through more doomed respawns.
-        crate::db_relaunch::note_respawn_failure(&app, &e).await;
-    }
-}
+mod db_wedge;
+pub use db_wedge::{
+    make_db_wedge_recovery_hook, new_db_wedge_breaker, DbWedgeBreaker, DbWedgeState,
+};
 
 #[derive(Clone, Debug)]
 pub(crate) struct InterruptedMeeting {
@@ -1576,109 +1342,9 @@ async fn kill_process_on_port(port: u16) {
 }
 
 #[cfg(test)]
-mod db_wedge_tests {
-    use super::{
-        capture_intended_now, db_wedge_recovery_decision, DbWedgeRecoveryDecision, DbWedgeState,
-        WedgeAction,
-    };
-    use screenpipe_db::WriteQueueHealth;
+mod capture_intent_tests {
+    use super::capture_intended_now;
     use std::sync::atomic::{AtomicBool, Ordering};
-    use std::time::{Duration, Instant};
-
-    const WINDOW: Duration = Duration::from_secs(600);
-    const MAX: usize = 3;
-
-    // First MAX signals restart; the next one gives up and notifies exactly
-    // once even though the breaker keeps being consulted.
-    #[test]
-    fn gives_up_after_cap_and_notifies_once() {
-        let mut s = DbWedgeState::default();
-        let t = Instant::now();
-        for _ in 0..MAX {
-            assert_eq!(s.decide(t, WINDOW, MAX), WedgeAction::Restart);
-        }
-        assert_eq!(
-            s.decide(t, WINDOW, MAX),
-            WedgeAction::GiveUp { notify: true }
-        );
-        // Hook keeps firing while tripped — no more notifications.
-        assert_eq!(
-            s.decide(t, WINDOW, MAX),
-            WedgeAction::GiveUp { notify: false }
-        );
-        assert_eq!(
-            s.decide(t, WINDOW, MAX),
-            WedgeAction::GiveUp { notify: false }
-        );
-    }
-
-    // Once the old restarts age out of the window, recovery re-arms: it restarts
-    // again and a fresh give-up re-notifies (it's a new corruption episode).
-    #[test]
-    fn restarts_age_out_and_re_arm_notification() {
-        let mut s = DbWedgeState::default();
-        let t0 = Instant::now();
-        for _ in 0..MAX {
-            assert_eq!(s.decide(t0, WINDOW, MAX), WedgeAction::Restart);
-        }
-        assert_eq!(
-            s.decide(t0, WINDOW, MAX),
-            WedgeAction::GiveUp { notify: true }
-        );
-
-        let later = t0 + WINDOW + Duration::from_secs(1);
-        for _ in 0..MAX {
-            assert_eq!(s.decide(later, WINDOW, MAX), WedgeAction::Restart);
-        }
-        assert_eq!(
-            s.decide(later, WINDOW, MAX),
-            WedgeAction::GiveUp { notify: true }
-        );
-    }
-
-    // A successful restart cadence (signals spaced beyond the window) never
-    // trips the breaker — every attempt restarts and nothing is suppressed.
-    #[test]
-    fn spaced_out_failures_never_trip() {
-        let mut s = DbWedgeState::default();
-        let mut t = Instant::now();
-        for _ in 0..10 {
-            assert_eq!(s.decide(t, WINDOW, MAX), WedgeAction::Restart);
-            t += WINDOW + Duration::from_secs(1);
-        }
-    }
-
-    #[test]
-    fn debounce_recheck_restarts_only_the_same_unrecovered_generation() {
-        let signaled = WriteQueueHealth::default();
-        let same_generation = signaled.clone();
-        let epoch = signaled.fatal_run_recovery_epoch();
-
-        assert_eq!(
-            db_wedge_recovery_decision(&signaled, epoch, Some(&same_generation)),
-            DbWedgeRecoveryDecision::Restart
-        );
-        assert_eq!(
-            db_wedge_recovery_decision(&signaled, epoch.wrapping_add(1), Some(&same_generation)),
-            DbWedgeRecoveryDecision::SkipRecovered
-        );
-    }
-
-    #[test]
-    fn debounce_recheck_ignores_stale_generation_signals() {
-        let signaled = WriteQueueHealth::default();
-        let replacement = WriteQueueHealth::default();
-        let epoch = signaled.fatal_run_recovery_epoch();
-
-        assert_eq!(
-            db_wedge_recovery_decision(&signaled, epoch, None),
-            DbWedgeRecoveryDecision::SkipNoServer
-        );
-        assert_eq!(
-            db_wedge_recovery_decision(&signaled, epoch, Some(&replacement)),
-            DbWedgeRecoveryDecision::SkipSupersededGeneration
-        );
-    }
 
     #[test]
     fn stop_during_debounce_is_honored_when_recovery_respawns() {
